refactor: IPNE 描画のワールド層を drawWorld へ分離
- 背景・マップ・パス・罠・壁・アイテム描画を drawWorld(frame, shakeOffset) へ逐語移植
- renderGameFrame は FrameContext を構築し同位置で drawWorld を呼ぶ
- getShakeOffset() の2重呼び出しを避けるため shakeOffset を引数として渡す設計を採用
- 特性化テストで描画コマンド列の不変を確認（2 passed / Snapshots: 2 passed）
